pixoo-api/requirements.txt
Pillow==9.2.0
Flask==2.1.3
flask-cors==3.0.10